pyproject.toml
[tool.poetry]
name = "bike-rental-backend"
version = "0.1.0"
description = ""
authors = ["Даниил Паутов <danpautoff@yandex.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
django = "^5.0.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
flake8 = "^7.1.0"
isort = "^5.13.2"
black = "^24.4.2"
pytest-django = "^4.8.0"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
model-bakery = "^1.18.2"

[tool.isort]
profile = "black"
skip_glob = ["**/migrations/*.py"]
src_paths = ["src/backend"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
